servers/requirements.txt
grpcio
grpcio-tools
fastapi
uvicorn
strawberry-graphql[fastapi]
psycopg2-binary
pgvector
python-dotenv
openai
requests